+++ src/components/Jokers/Audience/AudienceHelpJoker.module.css
@@ -1,0 +1,7 @@
+.audience.used{
+    opacity: 0.2;
+  }
+
+.audience {
+    background-image: url("assets/images/audience.jpg");
+}
