+++ src/components/Jokers/FiftyFifty/FiftyFiftyJoker.module.css
@@ -1,0 +1,7 @@
+.fifty {
+    background-image: url("assets/images/Classic5050.jpg");
+}
+/* this will be needed for the other jokers too */
+.fifty.used{
+    opacity: 0.2;
+  }
